+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix some weirdness of UI warnings
 * Add toggle all mods button
 * Fix SpireReturn memory leak
+* Fix crash if update save contains an error
 
 #### v3.10.1 ####
 * Fix crash in incomplete last-update code