+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add toggle all mods button
 * Fix SpireReturn memory leak
 * Fix crash if update save contains an error
+* Add mod profiles
 
 #### v3.10.1 ####
 * Fix crash in incomplete last-update code